rethink model, clean up logic for item & label apis

+++ server/api/label/labelApi.ts
@@ -8,9 +8,14 @@ const router = express.Router();
 
 router.use(authenticateUser);
 
-router.get("/", async (_req: Request, res: Response) => {
+router.get("/", async (req: Request, res: Response) => {
   try {
-    const labels = await prisma.label.findMany();
+    const labels = await prisma.label.findMany({
+      where: {
+        userId: (req as any).user.id,
+      },
+    });
+
     res.status(200).json(labels);
   } catch (error) {
     console.error("Error fetching labels:", error);
@@ -21,9 +26,16 @@ router.get("/", async (_req: Request, res: Response) => {
 router.get("/:id", async (req: Request<{ id: string }>, res: Response) => {
   try {
     const label = await prisma.label.findUnique({
-      where: { id: parseInt(req.params.id) },
+      where: { id: parseInt(req.params.id), userId: (req as any).user.id },
+      include: {
+        items: true,
+      },
     });
-    if (!label) return res.status(404).json({ message: "Label not found" });
+
+    if (!label) {
+      return res.status(404).json({ message: "Label not found for this user" });
+    }
+
     res.status(200).json(label);
   } catch (error) {
     console.error("Error fetching label:", error);
@@ -34,19 +46,39 @@ router.get("/:id", async (req: Request<{ id: string }>, res: Response) => {
 router.post(
   "/",
   async (req: Request<{}, {}, CreateLabelRequest>, res: Response) => {
+    if (req.body.itemIds) {
+      const items = await prisma.item.findMany({
+        where: {
+          id: {
+            in: req.body.itemIds,
+          },
+          userId: (req as any).user.id,
+        },
+      });
+      if (items.length !== req.body.itemIds.length) {
+        return res.status(400).json({
+          message:
+            "Invalid item IDs, ensure provided item IDs belong to the user",
+        });
+      }
+    }
+
     try {
       const label = await prisma.label.create({
-        data: req.body,
+        data: {
+          name: req.body.name,
+          colour: req.body.colour,
+          description: req.body.description,
+          userId: (req as any).user.id,
+          items: {
+            connect: req.body.itemIds?.map((id) => ({ id })),
+          },
+        },
       });
+
       res.status(201).json(label);
     } catch (error) {
       console.error("Error creating label:", error);
-      if (error instanceof Prisma.PrismaClientKnownRequestError) {
-        if (error.code === "P2002") {
-          res.status(400).json({ message: "Label name must be unique" });
-          return;
-        }
-      }
       res.status(500).json({ message: "Error creating label" });
     }
   }
@@ -59,23 +91,44 @@ router.put(
     res: Response
   ) => {
     try {
+      const existingLabel = await prisma.label.findUnique({
+        where: { id: parseInt(req.params.id), userId: (req as any).user.id },
+      });
+      if (!existingLabel) {
+        return res
+          .status(404)
+          .json({ message: "Label not found for this user" });
+      }
+      if (req.body.itemIds) {
+        const items = await prisma.item.findMany({
+          where: {
+            id: {
+              in: req.body.itemIds,
+            },
+            userId: (req as any).user.id,
+          },
+        });
+        if (items.length !== req.body.itemIds.length) {
+          return res.status(400).json({
+            message:
+              "Invalid item IDs, ensure provided item IDs belong to the user",
+          });
+        }
+      }
       const label = await prisma.label.update({
         where: { id: parseInt(req.params.id) },
-        data: req.body,
+        data: {
+          name: req.body.name,
+          colour: req.body.colour,
+          description: req.body.description,
+          items: {
+            connect: req.body.itemIds?.map((id) => ({ id })),
+          },
+        },
       });
       res.status(200).json(label);
     } catch (error) {
       console.error("Error updating label:", error);
-      if (error instanceof Prisma.PrismaClientKnownRequestError) {
-        if (error.code === "P2025") {
-          res.status(404).json({ message: "Label not found" });
-          return;
-        }
-        if (error.code === "P2002") {
-          res.status(400).json({ message: "Label name must be unique" });
-          return;
-        }
-      }
       res.status(500).json({ message: "Error updating label" });
     }
   }
@@ -83,22 +136,21 @@ router.put(
 
 router.delete("/:id", async (req: Request<{ id: string }>, res: Response) => {
   try {
+    const existingLabel = await prisma.label.findUnique({
+      where: { id: parseInt(req.params.id), userId: (req as any).user.id },
+    });
+    if (!existingLabel) {
+      return res.status(404).json({ message: "Label not found for this user" });
+    }
     await prisma.label.delete({
       where: { id: parseInt(req.params.id) },
     });
-    res.status(204).send();
+    res.status(200).send({
+      message: "Label deleted successfully",
+      label: existingLabel,
+    });
   } catch (error) {
     console.error("Error deleting label:", error);
-    if (error instanceof Prisma.PrismaClientKnownRequestError) {
-      if (error.code === "P2025") {
-        res.status(404).json({ message: "Label not found" });
-        return;
-      }
-      if (error.code === "P2003") {
-        res.status(400).json({ message: "Cannot delete label that is in use" });
-        return;
-      }
-    }
     res.status(500).json({ message: "Error deleting label" });
   }
 });
